+++ .nvmrc
@@ -1,0 +1,1 @@
+v23.11.0
