+++ packages/frontend/config/webpack/tsconfig.webpack.json
@@ -1,0 +1,17 @@
+
+{
+    "compilerOptions": {
+      "target": "ES6",
+      "module": "CommonJS",
+      "moduleResolution": "node",
+      "esModuleInterop": true,
+      "resolveJsonModule": true,
+      "sourceMap": true,
+      "strict": true,
+      "skipLibCheck": true,
+      "allowSyntheticDefaultImports": true,
+      "forceConsistentCasingInFileNames": true
+    },
+    "include": ["webpack.config.ts"],
+    "exclude": ["node_modules", "src"]
+  }
