+++ products/maistro-websites/frontend/package.json
@@ -32,6 +32,7 @@
     "@maistro/prettier-config": "workspace:^",
     "@maistro/typescript-config": "workspace:^",
     "@maistro/ui": "workspace:^",
+    "@maistro/webpack-config": "workspace:^",
     "@radix-ui/react-accordion": "^1.2.4",
     "@radix-ui/react-alert-dialog": "^1.1.7",
     "@radix-ui/react-aspect-ratio": "^1.1.3",
@@ -59,12 +60,12 @@
     "@radix-ui/react-toggle": "^1.1.3",
     "@radix-ui/react-toggle-group": "^1.1.3",
     "@radix-ui/react-tooltip": "^1.2.0",
-    "@typescript-eslint/parser": "^6.21.0",
     "@types/jest": "^29.5.3",
     "@types/node": "^22.14.1",
     "@types/react": "^19.0.12",
     "@types/react-dom": "^19.0.4",
     "@types/webpack": "^5.28.5",
+    "@typescript-eslint/parser": "^6.21.0",
     "autoprefixer": "^10.4.21",
     "css-loader": "^7.1.2",
     "eslint": "^9.23.0",
@@ -83,6 +84,7 @@
     "typescript": "^5.8.2",
     "webpack": "^5.91.0",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.1"
+    "webpack-dev-server": "^5.0.1",
+    "webpack-merge": "^6.0.1"
   }
 }
