+++ products/maistro-websites/frontend/webpack.config.ts
@@ -1,94 +1,17 @@
 import path from 'path';
 import { fileURLToPath } from 'url';
-import Webpack from 'webpack';
-import WebpackDevServer from 'webpack-dev-server';
-import HtmlWebpackPlugin from 'html-webpack-plugin';
-import MiniCssExtractPlugin from 'mini-css-extract-plugin';
-
-type WebpackConfiguration = Webpack.Configuration;
-type WebpackDevServerConfiguration = WebpackDevServer.Configuration
-interface Configuration extends WebpackConfiguration {
-  devServer?: WebpackDevServerConfiguration;
-}
+import {merge} from "webpack-merge"
+import createWebpackConfig from "@maistro/webpack-config"
 
 const isProduction = process.env.NODE_ENV === 'production';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-const config: Configuration = {
-  mode: isProduction ? 'production' : 'development',
-  entry: './src/index.tsx',
-  output: {
-    path: path.resolve(__dirname, 'dist'),
-    filename: '[name].[contenthash].js',
-    clean: true,
-    publicPath: '/',
-  },
-  devtool: isProduction ? 'source-map' : 'inline-source-map',
-  module: {
-    rules: [
-      {
-        test: /\.(ts|tsx)$/,
-        exclude: /node_modules/,
-        use: 'ts-loader',
-      },
-      {
-        test: /\.scss$/,
-        use: [
-          isProduction ? MiniCssExtractPlugin.loader : 'style-loader',
-          'css-loader',
-          'postcss-loader',
-          'sass-loader',
-        ],
-      },
-      {
-        test: /\.css$/,
-        use: [
-          isProduction ? MiniCssExtractPlugin.loader : 'style-loader',
-          'css-loader',
-          'postcss-loader',
-        ],
-      },
-      {
-        test: /\.(png|svg|jpg|jpeg|gif)$/i,
-        type: 'asset/resource',
-      },
-      {
-        test: /\.(woff|woff2|eot|ttf|otf)$/i,
-        type: 'asset/resource',
-      },
-    ],
-  },
-  resolve: {
-    extensions: ['.tsx', '.ts', '.js', '.jsx'],
-    alias: {
-      '@': path.resolve(__dirname, 'src'),
-    },
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: './public/index.html',
-      filename: 'index.html',
-    }),
-  ],
-  devServer: {
-    static: {
-      directory: path.join(__dirname, 'public'),
-    },
-    historyApiFallback: true,
-    port: 3000,
-    hot: true,
-    open: true,
-  },
-};
-
-if (isProduction) {
-  config.plugins?.push(
-    new MiniCssExtractPlugin({
-      filename: '[name].[contenthash].css',
-    })
-  );
+const config = () => {
+  return merge(
+    createWebpackConfig({isProduction, dirname: __dirname}),
+  )
 }
 
 export default config;
